Выделение общего кода ограничения запросов в трейт Limitable
