+++ src/Operation/TownSearchOperation.php
@@ -4,10 +4,13 @@ namespace Measoft\Operation;
 
 use Measoft\MeasoftException;
 use Measoft\Object\Town;
+use Measoft\Traits\Limitable;
 use SimpleXMLElement;
 
 class TownSearchOperation extends AbstractOperation
 {
+    use Limitable;
+
     /** @var string $zipCode Поиск по индексу */
     private $zipCode;
 
@@ -37,15 +40,6 @@ class TownSearchOperation extends AbstractOperation
 
     /** @var string $countryCode Поиск только по стране с указанным кодом */
     private $countryCode;
-
-    /** @var int $offset Задает номер записи результата, начиная с которой выдавать ответ */
-    private $offset = 0;
-
-    /** @var int $limit Задает количество записей результата, которые нужно вернуть */
-    private $limit = 10000;
-
-    /** @var string $countAll YES указывает на необходимость подсчета общего количества найденных совпадений */
-    private $countAll;
 
     /**
      * Поиск по индексу
@@ -178,45 +172,6 @@ class TownSearchOperation extends AbstractOperation
     }
 
     /**
-     * Задать номер записи результата, начиная с которой выдавать ответ
-     *
-     * @param int $offset Номер записи
-     * @return self
-     */
-    public function offset(int $offset): self
-    {
-        $this->offset = $offset;
-
-        return $this;
-    }
-
-    /**
-     * Задаеть количество записей результата, которые нужно вернуть
-     *
-     * @param int $limit Количество записей результата
-     * @return self
-     */
-    public function limit(int $limit): self
-    {
-        $this->limit = $limit;
-
-        return $this;
-    }
-
-    /**
-     * Установить подсчет общего количества найденных совпадений
-     *
-     * @param bool $countAll Вести подсчет общего количества найденных совпадений
-     * @return self
-     */
-    public function countAll(bool $countAll = true): self
-    {
-        $this->countAll = $countAll === true ? 'YES' : null;
-
-        return $this;
-    }
-
-    /**
      * Сформировать XML
      *
      * @return SimpleXMLElement
@@ -226,7 +181,6 @@ class TownSearchOperation extends AbstractOperation
         $xml        = $this->createXml('townlist');
         $codesearch = $xml->addChild('codesearch');
         $conditions = $xml->addChild('conditions');
-        $limit      = $xml->addChild('limit');
 
         $codesearch->addChild('zipcode', $this->zipCode);
         $codesearch->addChild('kladrcode', $this->kladrCode);
@@ -240,9 +194,7 @@ class TownSearchOperation extends AbstractOperation
         $conditions->addChild('fullname', $this->fullName);
         $conditions->addChild('country', $this->countryCode);
 
-        $limit->addChild('limitfrom', $this->offset);
-        $limit->addChild('limitcount', $this->limit);
-        $limit->addChild('countall', $this->countAll);
+        $this->buildLimitXML($xml);
 
         return $xml;
     }
